requirements.txt
pytesseract
PyPDF2
pillow
speechrecognition
pydub
requests
pandas
sqlite3
